pyproject.toml
[tool.poetry]
name = "tms-django"
version = "0.1.0"
description = ""
authors = ["elya-hp <elya.hp@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
djangorestframework = "^3.15.1"
httpx = "^0.27.0"
drf-spectacular = "^0.27.2"
loguru = "^0.7.2"
pre-commit = "^3.7.0"
factory-boy = "^3.3.0"
faker = "^24.9.0"
django-filter = "^24.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
